refactor(api): add BadgeCriterionForCalculation specific model file
