+++ api/tests/tooling/domain-builder/factory/build-badge-criterion-for-calculation.js
@@ -1,10 +1,10 @@
-import { BadgeCriterion } from '../../../../lib/domain/models/BadgeForCalculation.js';
+import { BadgeCriterionForCalculation } from '../../../../lib/domain/models/BadgeCriterionForCalculation.js';
 
 const buildBadgeCriterionForCalculation = function buildBadgeCriterionForCalculation({
   threshold = 80,
   skillIds = ['recCoucou'],
 } = {}) {
-  return new BadgeCriterion({
+  return new BadgeCriterionForCalculation({
     threshold,
     skillIds,
   });
